started styling Product component. Added add to card button, which needs functionality
