+++ client/src/components/Product/style.css
@@ -1,0 +1,10 @@
+.product {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 20%;
+}
+
+@media screen and (min-width: 480px) {
+
+  }
